Rename mir_cfg.rs to emit_tir.rs.
The old name was a throwback to when we though we would simply serialise
MIR, instead of designing our own tracing byte-code.
